Development (#24)
* Improve KpiService code

* Add logging info

* Comment locks logging

* Add logging info

* Sync attempt

* MES-128 add stack tracing

* Feature/mc 231 (#21)

* MES-231 - Laboratory: endpoint create composed production order

* MES-231 - Laboratory: endpoint create composed production order

* MES-231 - Laboratory: endpoint create composed production order

* MES-231 - Laboratory: endpoint create composed production order

* MES-231 - Laboratory: endpoint create composed production order

* MES-231 - Laboratory: endpoint create composed production order

* MES-231 - Laboratory: endpoint create composed production order

* MES-231 - Laboratory: endpoint create composed production order

* MES-231 - Laboratory: endpoint create composed production order

* MES-236 - Laboratory: create sample creation logic (#22)

* MES-236 - Laboratory: create sample creation logic

* MES-236 - Laboratory: create sample creation logic

* MES-236 - Laboratory: create sample creation logic

* MEs-128 - IMS basic implementation

* MES-128 - add IMS to countingEquipmentDto

* Feature/mc 233 (#23)

* MES-233 - Laboratory: endpoint create hits

* MES-233 - Laboratory: endpoint create hits

* MES-233 - Laboratory: endpoint create hits

* MES-233 - Laboratory: endpoint create hits

* MES-233 - Laboratory: endpoint create hits

* MES-233 - Laboratory: endpoint create hits

* MES-233 - Laboratory: endpoint create hits

* MES-233 - Laboratory: endpoint create hits

* MES-128 - update postman collection

* MES-128 - Update equipment IMS

* feature/MES-128 small typo fix

* MES-128 - add IMS validations

* MES-128 - remove TODO

* MES-128 - minor refactor

* MEs-128 - add ims to production orders

---------

+++ src/main/java/com/tde/mescloud/model/entity/SampleEntity.java
@@ -24,9 +24,9 @@ public class SampleEntity {
 
     private int amount;
 
-    private double tcaAverage;
+    private int tcaAverage;
 
-    private double reliability;
+    private int reliability;
 
     @CreationTimestamp
     @Temporal(TemporalType.TIMESTAMP)
